backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.1.8
colorzero==2.0
exceptiongroup==1.3.0
fastapi==0.117.1
gpiozero==2.0.1
h11==0.16.0
idna==3.10
importlib-metadata==4.13.0
importlib-resources==5.13.0
lgpio==0.2.2.0
pydantic==2.11.9
pydantic_core==2.33.2
RPi.GPIO==0.7.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.36.0
zipp==3.23.0
